@@ -3,6 +3,7 @@
   "version": "1.6.1",
   "description": "Super light-weight stand-in replacement for console logging to only output JSON",
   "main": "index.js",
+  "type": "module",
   "scripts": {
     "test": "jest",
     "lint": "eslint index.js"
@@ -27,16 +28,10 @@
   },
   "homepage": "https://github.com/sealninja/json-console-logger",
   "devDependencies": {
-    "eslint": "9.10.0",
-    "@eslint/compat": "1.1.1",
-    "@eslint/eslintrc": "3.1.0",
-    "@eslint/js": "9.9.1",
-    "eslint-config-airbnb": "19.0.4",
-    "eslint-config-prettier": "9.1.0",
-    "eslint-plugin-import": "2.30.0",
-    "eslint-plugin-jsx-a11y": "6.10.0",
-    "eslint-plugin-prettier": "5.2.1",
-    "eslint-plugin-react": "7.35.2",
+    "@dr.pogodin/eslint-configs": "0.0.8",
+    "eslint": "9.30.1",
+    "eslint-config-prettier": "^10.1.5",
+    "eslint-plugin-prettier": "^5.5.1",
     "jest": "29.7.0",
     "prettier": "3.3.3"
   }
